requirements.txt
Django==3.2
gunicorn==20.0.4
requests==2.25.1
pytest==6.2.3
pytest-django==4.2.0
agavepy==0.9.5
pytest-mock==3.5.1
pyjwt[crypto]==1.7.1
tapipy
matplotlib
